+++ src/psyclone/gocean0p1.py
@@ -40,8 +40,9 @@
 
 from __future__ import absolute_import
 from psyclone.configuration import Config
-from psyclone.psyGen import PSy, Invokes, Invoke, InvokeSchedule, Loop, \
-    CodedKern, Arguments, KernelArgument, Literal, Schedule
+from psyclone.psyir.nodes import Loop, Literal, Schedule
+from psyclone.psyGen import PSy, Invokes, Invoke, InvokeSchedule, \
+    CodedKern, Arguments, KernelArgument
 from psyclone.psyir.symbols import DataType
 from psyclone.parse.kernel import KernelType, Descriptor
 from psyclone.parse.utils import ParseError
@@ -259,7 +260,7 @@ class GOLoop(Loop):
 
         if self.field_space == "every":
             from psyclone.f2pygen import DeclGen
-            from psyclone.psyGen import BinaryOperation, Reference
+            from psyclone.psyir.nodes import BinaryOperation, Reference
             dim_var = DeclGen(parent, datatype="INTEGER",
                               entity_decls=[self._variable_name])
             parent.add(dim_var)
