+++ src/psyclone/psyGen.py
@@ -40,7 +40,6 @@
     particular API and implementation. '''
 
 from __future__ import print_function, absolute_import
-from enum import Enum
 import abc
 import re
 import six
@@ -49,28 +48,7 @@ from psyclone.configuration import Config
 from psyclone.f2pygen import DirectiveGen
 from psyclone.core.access_info import VariablesAccessInfo, AccessType
 from psyclone.psyir.symbols import SymbolTable, SymbolError, DataType
-
-# We use the termcolor module (if available) to enable us to produce
-# coloured, textual representations of Invoke schedules. If it's not
-# available then we don't use colour.
-try:
-    from termcolor import colored
-except ImportError:
-    # We don't have the termcolor package available so provide
-    # alternative routine
-    def colored(text, _):
-        '''
-        Returns the supplied text argument unchanged. This is a swap-in
-        replacement for when termcolor.colored is not available.
-
-        :param text: Text to return
-        :type text: string
-        :param _: Fake argument, only required to match interface
-                  provided by termcolor.colored
-        :returns: The supplied text, unchanged
-        :rtype: string
-        '''
-        return text
+from psyclone.psyir.nodes import Node, Schedule, Loop
 
 
 # The types of 'intent' that an argument to a Fortran subroutine
@@ -92,33 +70,6 @@ VALID_ARG_TYPE_NAMES = []
 
 # Mapping of access type to operator.
 REDUCTION_OPERATOR_MAPPING = {AccessType.SUM: "+"}
-
-# Colour map to use when writing Invoke schedule to terminal. (Requires
-# that the termcolor package be installed. If it isn't then output is not
-# coloured.) See https://pypi.python.org/pypi/termcolor for details.
-SCHEDULE_COLOUR_MAP = {"Schedule": "white",
-                       "Loop": "red",
-                       "GlobalSum": "cyan",
-                       "Directive": "green",
-                       "HaloExchange": "blue",
-                       "HaloExchangeStart": "yellow",
-                       "HaloExchangeEnd": "yellow",
-                       "BuiltIn": "magenta",
-                       "CodedKern": "magenta",
-                       "InlinedKern": "magenta",
-                       "Profile": "green",
-                       "Extract": "green",
-                       "If": "red",
-                       "Assignment": "blue",
-                       "Reference": "yellow",
-                       "Operation": "blue",
-                       "Literal": "yellow",
-                       "Return": "yellow",
-                       "CodeBlock": "red",
-                       "Container": "green"}
-
-# Default indentation string
-INDENTATION_STRING = "    "
 
 
 def object_index(alist, item):
@@ -948,846 +899,6 @@ class Invoke(object):
         # create the subroutine kernel call content
         self.schedule.gen_code(invoke_sub)
         parent.add(invoke_sub)
-
-
-class Node(object):
-    '''
-    Base class for a node in the PSyIR (schedule).
-
-    :param ast: reference into the fparser2 AST corresponding to this node.
-    :type ast: sub-class of :py:class:`fparser.two.Fortran2003.Base`
-    :param children: the PSyIR nodes that are children of this node.
-    :type children: list of :py:class:`psyclone.psyGen.Node`
-    :param parent: that parent of this node in the PSyIR tree.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-    :param annotations: Tags that provide additional information about \
-        the node. The node should still be functionally correct when \
-        ignoring these tags.
-    :type annotations: list of str
-
-    :raises InternalError: if an invalid annotation tag is supplied.
-
-    '''
-    # Define two class constants: START_DEPTH and START_POSITION
-    # START_DEPTH is used to calculate depth of all Nodes in the tree
-    # (1 for main Nodes and increasing for their descendants).
-    START_DEPTH = 0
-    # START_POSITION is used to to calculate position of all Nodes in
-    # the tree (absolute or relative to a parent).
-    START_POSITION = 0
-    # The list of valid annotations for this Node. Populated by sub-class.
-    valid_annotations = tuple()
-
-    def __init__(self, ast=None, children=None, parent=None, annotations=None):
-        if not children:
-            self._children = []
-        else:
-            self._children = children
-        self._parent = parent
-        # Reference into fparser2 AST (if any)
-        self._ast = ast
-        # Ref. to last fparser2 parse tree node associated with this Node.
-        # This is required when adding directives.
-        self._ast_end = None
-        # List of tags that provide additional information about this Node.
-        self._annotations = []
-        if annotations:
-            for annotation in annotations:
-                if annotation in self.valid_annotations:
-                    self._annotations.append(annotation)
-                else:
-                    raise InternalError(
-                        "{0} with unrecognized annotation '{1}', valid "
-                        "annotations are: {2}.".format(
-                            self.__class__.__name__, annotation,
-                            self.valid_annotations))
-        # Name to use for this Node type. By default we use the name of
-        # the class but this can be overridden by a sub-class.
-        self._text_name = self.__class__.__name__
-        # Which colour to use from the SCHEDULE_COLOUR_MAP
-        self._colour_key = self.__class__.__name__
-
-    def coloured_name(self, colour=True):
-        '''
-        Returns the display name of this Node, optionally with colour control
-        codes (requires that the termcolor package be installed).
-
-        :param bool colour: whether or not to include colour control codes \
-                            in the result.
-
-        :returns: the name of this node, optionally with colour control codes.
-        :rtype: str
-        '''
-        if colour:
-            try:
-                return colored(self._text_name,
-                               SCHEDULE_COLOUR_MAP[self._colour_key])
-            except KeyError:
-                pass
-        return self._text_name
-
-    def node_str(self, colour=True):
-        '''
-        :param bool colour: whether or not to include control codes for \
-                            coloured text.
-
-        :returns: a text description of this node. Will typically be \
-                  overridden by sub-class.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[]"
-
-    @abc.abstractmethod
-    def __str__(self):
-        return self.node_str(False)
-
-    def math_equal(self, other):
-        '''Returns True if the self has the same results as other. The
-        implementation in the base class just confirms that the type is the
-        same, and the number of children as well.
-
-        :param other: the node to compare self with.
-        :type other: py:class:`psyclone.psyGen.Node`.
-
-        :returns: whether self has the same result as other.
-        :rtype: bool
-        '''
-
-        # pylint: disable=unidiomatic-typecheck
-        if type(self) != type(other):
-            return False
-
-        if len(self.children) != len(other.children):
-            return False
-
-        for i, entity in enumerate(self.children):
-            if not entity.math_equal(other.children[i]):
-                return False
-        return True
-
-    @property
-    def ast(self):
-        '''
-        :returns: a reference to that part of the fparser2 parse tree that \
-                  this node represents or None.
-        :rtype: sub-class of :py:class:`fparser.two.utils.Base`
-        '''
-        return self._ast
-
-    @property
-    def ast_end(self):
-        '''
-        :returns: a reference to the last node in the fparser2 parse tree \
-                  that represents a child of this PSyIR node or None.
-        :rtype: sub-class of :py:class:`fparser.two.utils.Base`
-        '''
-        return self._ast_end
-
-    @ast.setter
-    def ast(self, ast):
-        '''
-        Set a reference to the fparser2 node associated with this Node.
-
-        :param ast: fparser2 node associated with this Node.
-        :type ast: :py:class:`fparser.two.utils.Base`
-        '''
-        self._ast = ast
-
-    @ast_end.setter
-    def ast_end(self, ast_end):
-        '''
-        Set a reference to the last fparser2 node associated with this Node.
-
-        :param ast: last fparser2 node associated with this Node.
-        :type ast: :py:class:`fparser.two.utils.Base`
-        '''
-        self._ast_end = ast_end
-
-    @property
-    def annotations(self):
-        ''' Return the list of annotations attached to this Node.
-
-        :returns: List of anotations
-        :rtype: list of str
-        '''
-        return self._annotations
-
-    def dag(self, file_name='dag', file_format='svg'):
-        '''Create a dag of this node and its children.'''
-        try:
-            import graphviz as gv
-        except ImportError:
-            # todo: add a warning to a log file here
-            # silently return if graphviz bindings are not installed
-            return
-        try:
-            graph = gv.Digraph(format=file_format)
-        except ValueError:
-            raise GenerationError(
-                "unsupported graphviz file format '{0}' provided".
-                format(file_format))
-        self.dag_gen(graph)
-        graph.render(filename=file_name)
-
-    def dag_gen(self, graph):
-        '''Output my node's graph (dag) information and call any
-        children. Nodes with children are represented as two vertices,
-        a start and an end. Forward dependencies are represented as
-        green edges, backward dependencies are represented as red
-        edges (but their direction is reversed so the layout looks
-        reasonable) and parent child dependencies are represented as
-        blue edges.'''
-        # names to append to my default name to create start and end vertices
-        start_postfix = "_start"
-        end_postfix = "_end"
-        if self.children:
-            # I am represented by two vertices, a start and an end
-            graph.node(self.dag_name+start_postfix)
-            graph.node(self.dag_name+end_postfix)
-        else:
-            # I am represented by a single vertex
-            graph.node(self.dag_name)
-        # first deal with forward dependencies
-        remote_node = self.forward_dependence()
-        local_name = self.dag_name
-        if self.children:
-            # edge will come from my end vertex as I am a forward dependence
-            local_name += end_postfix
-        if remote_node:
-            # this node has a forward dependence
-            remote_name = remote_node.dag_name
-            if remote_node.children:
-                # the remote node has children so I will connect to
-                # its start vertex
-                remote_name += start_postfix
-            # Create the forward dependence edge in green
-            graph.edge(local_name, remote_name, color="green")
-        elif self.parent:
-            # this node is a child of another node and has no forward
-            # dependence. Therefore connect it to the the end vertex
-            # of its parent. Use blue to indicate a parent child
-            # relationship.
-            remote_name = self.parent.dag_name + end_postfix
-            graph.edge(local_name, remote_name, color="blue")
-        # now deal with backward dependencies. When creating the edges
-        # we reverse the direction of the dependence (place
-        # remote_node before local_node) to help with the graph
-        # layout
-        remote_node = self.backward_dependence()
-        local_name = self.dag_name
-        if self.children:
-            # the edge will come from my start vertex as I am a
-            # backward dependence
-            local_name += start_postfix
-        if remote_node:
-            # this node has a backward dependence.
-            remote_name = remote_node.dag_name
-            if remote_node.children:
-                # the remote node has children so I will connect to
-                # its end vertex
-                remote_name += end_postfix
-            # Create the backward dependence edge in red.
-            graph.edge(remote_name, local_name, color="red")
-        elif self.parent:
-            # this node has a parent and has no backward
-            # dependence. Therefore connect it to the the start vertex
-            # of its parent. Use blue to indicate a parent child
-            # relationship.
-            remote_name = self.parent.dag_name + start_postfix
-            graph.edge(remote_name, local_name, color="blue")
-        # now call any children so they can add their information to
-        # the graph
-        if isinstance(self, Loop):
-            # In case of a loop only loop at the body (the other part
-            # of the tree contain start, stop, step values):
-            self.loop_body.dag_gen(graph)
-        else:
-            for child in self.children:
-                child.dag_gen(graph)
-
-    @property
-    def dag_name(self):
-        '''Return the base dag name for this node.'''
-        return "node_" + str(self.abs_position)
-
-    @property
-    def args(self):
-        '''Return the list of arguments associated with this Node. The default
-        implementation assumes the Node has no directly associated
-        arguments (i.e. is not a Kern class or subclass). Arguments of
-        any of this nodes descendants are considered to be
-        associated. '''
-        args = []
-        for call in self.kernels():
-            args.extend(call.args)
-        return args
-
-    def backward_dependence(self):
-        '''Returns the closest preceding Node that this Node has a direct
-        dependence with or None if there is not one. Only Nodes with
-        the same parent as self are returned. Nodes inherit their
-        descendants' dependencies. The reason for this is that for
-        correctness a node must maintain its parent if it is
-        moved. For example a halo exchange and a kernel call may have
-        a dependence between them but it is the loop body containing
-        the kernel call that the halo exchange must not move beyond
-        i.e. the loop body inherits the dependencies of the routines
-        within it.'''
-        dependence = None
-        # look through all the backward dependencies of my arguments
-        for arg in self.args:
-            dependent_arg = arg.backward_dependence()
-            if dependent_arg:
-                # this argument has a backward dependence
-                node = dependent_arg.call
-                # if the remote node is deeper in the tree than me
-                # then find the ancestor that is at the same level of
-                # the tree as me.
-                while node.depth > self.depth:
-                    node = node.parent
-                if self.sameParent(node):
-                    # The remote node (or one of its ancestors) shares
-                    # the same parent as me
-                    if not dependence:
-                        # this is the first dependence found so keep it
-                        dependence = node
-                    else:
-                        # we have already found a dependence
-                        if dependence.position < node.position:
-                            # the new dependence is closer to me than
-                            # the previous dependence so keep it
-                            dependence = node
-        return dependence
-
-    def forward_dependence(self):
-        '''Returns the closest following Node that this Node has a direct
-        dependence with or None if there is not one. Only Nodes with
-        the same parent as self are returned. Nodes inherit their
-        descendants' dependencies. The reason for this is that for
-        correctness a node must maintain its parent if it is
-        moved. For example a halo exchange and a kernel call may have
-        a dependence between them but it is the loop body containing
-        the kernel call that the halo exchange must not move beyond
-        i.e. the loop body inherits the dependencies of the routines
-        within it.'''
-        dependence = None
-        # look through all the forward dependencies of my arguments
-        for arg in self.args:
-            dependent_arg = arg.forward_dependence()
-            if dependent_arg:
-                # this argument has a forward dependence
-                node = dependent_arg.call
-                # if the remote node is deeper in the tree than me
-                # then find the ancestor that is at the same level of
-                # the tree as me.
-                while node.depth > self.depth:
-                    node = node.parent
-                if self.sameParent(node):
-                    # The remote node (or one of its ancestors) shares
-                    # the same parent as me
-                    if not dependence:
-                        # this is the first dependence found so keep it
-                        dependence = node
-                    else:
-                        if dependence.position > node.position:
-                            # the new dependence is closer to me than
-                            # the previous dependence so keep it
-                            dependence = node
-        return dependence
-
-    def is_valid_location(self, new_node, position="before"):
-        '''If this Node can be moved to the new_node
-        (where position determines whether it is before of after the
-        new_node) without breaking any data dependencies then return True,
-        otherwise return False.
-
-        :param new_node: Node to which this node should be moved.
-        :type new_node: :py:class:`psyclone.psyGen.Node`
-        :param str position: either 'before' or 'after'.
-
-        :raises GenerationError: if new_node is not an\
-                instance of :py:class:`psyclone.psyGen.Node`.
-        :raises GenerationError: if position is not 'before' or 'after'.
-        :raises GenerationError: if self and new_node do not have the same\
-                parent.
-        :raises GenerationError: self and new_node are the same Node.
-
-        :returns: whether or not the specified location is valid for this node.
-        :rtype: bool
-
-        '''
-        # First perform correctness checks
-        # 1: check new_node is a Node
-        if not isinstance(new_node, Node):
-            raise GenerationError(
-                "In the psyGen.Node.is_valid_location() method the "
-                "supplied argument is not a Node, it is a '{0}'.".
-                format(type(new_node).__name__))
-
-        # 2: check position has a valid value
-        valid_positions = ["before", "after"]
-        if position not in valid_positions:
-            raise GenerationError(
-                "The position argument in the psyGenNode.is_valid_location() "
-                "method must be one of {0} but found '{1}'".format(
-                    valid_positions, position))
-
-        # 3: check self and new_node have the same parent
-        if not self.sameParent(new_node):
-            raise GenerationError(
-                "In the psyGen.Node.is_valid_location() method "
-                "the node and the location do not have the same parent")
-
-        # 4: check proposed new position is not the same as current position
-        new_position = new_node.position
-        if new_position < self.position and position == "after":
-            new_position += 1
-        elif new_position > self.position and position == "before":
-            new_position -= 1
-
-        if self.position == new_position:
-            raise GenerationError(
-                "In the psyGen.Node.is_valid_location() method, the "
-                "node and the location are the same so this transformation "
-                "would have no effect.")
-
-        # Now determine whether the new location is valid in terms of
-        # data dependencies
-        # Treat forward and backward dependencies separately
-        if new_position < self.position:
-            # the new_node is before this node in the schedule
-            prev_dep_node = self.backward_dependence()
-            if not prev_dep_node:
-                # There are no backward dependencies so the move is valid
-                return True
-            else:
-                # return (is the dependent node before the new_position?)
-                return prev_dep_node.position < new_position
-        else:  # new_node.position > self.position
-            # the new_node is after this node in the schedule
-            next_dep_node = self.forward_dependence()
-            if not next_dep_node:
-                # There are no forward dependencies so the move is valid
-                return True
-            else:
-                # return (is the dependent node after the new_position?)
-                return next_dep_node.position > new_position
-
-    @property
-    def depth(self):
-        '''
-        Returns this Node's depth in the tree: 1 for the Schedule
-        and increasing for its descendants at each level.
-        :returns: depth of the Node in the tree
-        :rtype: int
-        '''
-        my_depth = self.START_DEPTH
-        node = self
-        while node is not None:
-            node = node.parent
-            my_depth += 1
-        return my_depth
-
-    def view(self, indent=0, index=None):
-        ''' Print out description of current node to stdout and
-        then call view() on all child nodes.
-
-        :param int indent: depth of indent for output text.
-        :param int index: the position of this Node wrt its siblings or None.
-
-        '''
-        if not isinstance(self.parent, Schedule) or index is None:
-            print("{0}{1}".format(self.indent(indent),
-                                  self.node_str(colour=True)))
-        else:
-            print("{0}{1}: {2}".format(self.indent(indent), index,
-                                       self.node_str(colour=True)))
-        for idx, entity in enumerate(self._children):
-            entity.view(indent=indent + 1, index=idx)
-
-    @staticmethod
-    def indent(count, indent=INDENTATION_STRING):
-        '''
-        Helper function to produce indentation strings.
-
-        :param int count: Number of indentation levels.
-        :param str indent: String representing one indentation level.
-        :returns: Complete indentation string.
-        :rtype: str
-        '''
-        return count * indent
-
-    def list(self, indent=0):
-        result = ""
-        for entity in self._children:
-            result += str(entity)+"\n"
-        return result
-
-    def addchild(self, child, index=None):
-        if index is not None:
-            self._children.insert(index, child)
-        else:
-            self._children.append(child)
-
-    @property
-    def children(self):
-        return self._children
-
-    @children.setter
-    def children(self, my_children):
-        self._children = my_children
-
-    @property
-    def parent(self):
-        return self._parent
-
-    @parent.setter
-    def parent(self, my_parent):
-        self._parent = my_parent
-
-    @property
-    def position(self):
-        '''
-        Find a Node's position relative to its parent Node (starting
-        with 0 if it does not have a parent).
-
-        :returns: relative position of a Node to its parent
-        :rtype: int
-        '''
-        if self.parent is None:
-            return self.START_POSITION
-        return self.parent.children.index(self)
-
-    @property
-    def abs_position(self):
-        '''
-        Find a Node's absolute position in the tree (starting with 0 if
-        it is the root). Needs to be computed dynamically from the
-        starting position (0) as its position may change.
-
-        :returns: absolute position of a Node in the tree
-        :rtype: int
-
-        :raises InternalError: if the absolute position cannot be found
-        '''
-        if self.root == self and isinstance(self.root, Schedule):
-            return self.START_POSITION
-        found, position = self._find_position(self.root.children,
-                                              self.START_POSITION)
-        if not found:
-            raise InternalError("Error in search for Node position "
-                                "in the tree")
-        return position
-
-    def _find_position(self, children, position):
-        '''
-        Recurse through the tree depth first returning position of
-        a Node if found.
-        :param children: list of Nodes which are children of this Node
-        :type children: list of :py:class:`psyclone.psyGen.Node`
-        :returns: position of the Node in the tree
-        :rtype: int
-        :raises InternalError: if the starting position is < 0
-        '''
-        if position < self.START_POSITION:
-            raise InternalError(
-                "Search for Node position started from {0} "
-                "instead of {1}.".format(position, self.START_POSITION))
-        for child in children:
-            position += 1
-            if child == self:
-                return True, position
-            if child.children:
-                found, position = self._find_position(child.children, position)
-                if found:
-                    return True, position
-        return False, position
-
-    @property
-    def root(self):
-        node = self
-        while node.parent is not None:
-            node = node.parent
-        return node
-
-    def sameRoot(self, node_2):
-        if self.root == node_2.root:
-            return True
-        return False
-
-    def sameParent(self, node_2):
-        if self.parent is None or node_2.parent is None:
-            return False
-        if self.parent == node_2.parent:
-            return True
-        return False
-
-    def walk(self, my_type, stop_type=None):
-        ''' Recurse through the PSyIR tree and return all objects that are
-        an instance of 'my_type', which is either a single class or a tuple
-        of classes. In the latter case all nodes are returned that are
-        instances of any classes in the tuple. The recursion into the tree
-        is stopped if an instance of 'stop_type' (which is either a single
-        class or a tuple of classes) is found. This can be used to avoid
-        analysing e.g. inlined kernels, or as performance optimisation to
-        reduce the number of recursive calls.
-
-        :param my_type: the class(es) for which the instances are collected.
-        :type my_type: either a single :py:class:`psyclone.Node` class \
-            or a tuple of such classes
-        :param stop_type: class(es) at which recursion is halted (optional)."
-
-        :type stop_type: None or a single :py:class:`psyclone.Node` \
-            class or a tuple of such classes
-
-        :returns: list with all nodes that are instances of my_type \
-            starting at and including this node.
-        :rtype: list of :py:class:`psyclone.Node` instances.
-        '''
-        local_list = []
-        if isinstance(self, my_type):
-            local_list.append(self)
-
-        # Stop recursion further into the tree if an instance of a class
-        # listed in stop_type is found.
-        if stop_type and isinstance(self, stop_type):
-            return local_list
-        for child in self.children:
-            local_list += child.walk(my_type, stop_type)
-        return local_list
-
-    def ancestor(self, my_type, excluding=None, include_self=False):
-        '''
-        Search back up tree and check whether we have an ancestor that is
-        an instance of the supplied type. If we do then we return
-        it otherwise we return None. An individual (or tuple of) (sub-)
-        class(es) to ignore may be provided via the `excluding` argument. If
-        include_self is True then the current node is included in the search.
-
-        :param my_type: class(es) to search for.
-        :type my_type: type or tuple of types
-        :param tuple excluding: individual (or tuple of) (sub-)class(es) to \
-                                ignore or None.
-        :param bool include_self: whether or not to include this node in the \
-                                  search.
-
-        :returns: First ancestor Node that is an instance of any of the \
-                  requested classes or None if not found.
-        :rtype: :py:class:`psyclone.psyGen.Node` or NoneType
-
-        :raises TypeError: if `excluding` is provided but is not a type or \
-                           tuple of types.
-        '''
-        if include_self:
-            myparent = self
-        else:
-            myparent = self.parent
-
-        if excluding is not None:
-            if isinstance(excluding, type):
-                excludes = (excluding, )
-            elif isinstance(excluding, tuple):
-                excludes = excluding
-            else:
-                raise TypeError(
-                    "The 'excluding' argument to ancestor() must be a type or "
-                    "a tuple of types but got: '{0}'".format(
-                        type(excluding).__name__))
-
-        while myparent is not None:
-            if isinstance(myparent, my_type):
-                if not (excluding and isinstance(myparent, excludes)):
-                    # This parent node is not an instance of an excluded
-                    # sub-class so return it
-                    return myparent
-            myparent = myparent.parent
-        return None
-
-    def kernels(self):
-        '''
-        :returns: all kernels that are descendants of this node in the PSyIR.
-        :rtype: list of :py:class:`psyclone.psyGen.Kern` sub-classes.
-        '''
-        return self.walk(Kern)
-
-    def following(self):
-        '''Return all :py:class:`psyclone.psyGen.Node` nodes after me in the
-        schedule. Ordering is depth first.
-
-        :returns: a list of nodes
-        :rtype: :func:`list` of :py:class:`psyclone.psyGen.Node`
-
-        '''
-        all_nodes = self.root.walk(Node)
-        position = all_nodes.index(self)
-        return all_nodes[position+1:]
-
-    def preceding(self, reverse=None):
-        '''Return all :py:class:`psyclone.psyGen.Node` nodes before me in the
-        schedule. Ordering is depth first. If the `reverse` argument
-        is set to `True` then the node ordering is reversed
-        i.e. returning the nodes closest to me first
-
-        :param: reverse: An optional, default `False`, boolean flag
-        :type: reverse: bool
-        :returns: A list of nodes
-        :rtype: :func:`list` of :py:class:`psyclone.psyGen.Node`
-
-        '''
-        all_nodes = self.root.walk(Node)
-        position = all_nodes.index(self)
-        nodes = all_nodes[:position]
-        if reverse:
-            nodes.reverse()
-        return nodes
-
-    def coded_kernels(self):
-        '''
-        Returns a list of all of the user-supplied kernels (as opposed to
-        builtins) that are beneath this node in the PSyIR.
-
-        :returns: all user-supplied kernel calls below this node.
-        :rtype: list of :py:class:`psyclone.psyGen.CodedKern`
-        '''
-        return self.walk(CodedKern)
-
-    def loops(self):
-        '''Return all loops currently in this schedule.'''
-        return self.walk(Loop)
-
-    def reductions(self, reprod=None):
-        '''Return all calls that have reductions and are decendents of this
-        node. If reprod is not provided, all reductions are
-        returned. If reprod is False, all builtin reductions that are
-        not set to reproducible are returned. If reprod is True, all
-        builtins that are set to reproducible are returned.'''
-
-        call_reduction_list = []
-        for call in self.walk(Kern):
-            if call.is_reduction:
-                if reprod is None:
-                    call_reduction_list.append(call)
-                elif reprod:
-                    if call.reprod_reduction:
-                        call_reduction_list.append(call)
-                else:
-                    if not call.reprod_reduction:
-                        call_reduction_list.append(call)
-        return call_reduction_list
-
-    def is_openmp_parallel(self):
-        ''':returns: True if this Node is within an OpenMP parallel region.
-
-        '''
-        omp_dir = self.ancestor(OMPParallelDirective)
-        if omp_dir:
-            return True
-        return False
-
-    def gen_code(self, parent):
-        '''Abstract base class for code generation function.
-
-        :param parent: the parent of this Node in the PSyIR.
-        :type parent: :py:class:`psyclone.psyGen.Node`
-        '''
-        raise NotImplementedError("Please implement me")
-
-    def update(self):
-        ''' By default we assume there is no need to update the existing
-        fparser2 AST which this Node represents. We simply call the update()
-        method of any children. '''
-        for child in self._children:
-            child.update()
-
-    def reference_accesses(self, var_accesses):
-        '''Get all variable access information. The default implementation
-        just recurses down to all children.
-
-        :param var_accesses: Stores the output results.
-        :type var_accesses: \
-            :py:class:`psyclone.core.access_info.VariablesAccessInfo`
-        '''
-        for child in self._children:
-            child.reference_accesses(var_accesses)
-
-    def _insert_schedule(self, children=None, ast=None):
-        '''
-        Utility method to insert a Schedule between this Node and the
-        supplied list of children.
-
-        :param children: nodes which will become children of the \
-                         new Schedule.
-        :type children: list of :py:class:`psyclone.psyGen.Node`
-        :param ast: reference to fparser2 parse tree for associated \
-                    Fortran code.
-        :type ast: :py:class:`fparser.two.utils.Base`
-
-        :returns: the new Schedule node.
-        :rtype: :py:class:`psyclone.psyGen.Schedule`
-        '''
-        sched = Schedule(children=children, parent=self)
-        if children:
-            # If we have children then set the Schedule's AST pointer to
-            # point to the AST associated with them.
-            sched.ast = children[0].ast
-            for child in children:
-                child.parent = sched
-        else:
-            sched.ast = ast
-        return sched
-
-
-class Schedule(Node):
-    ''' Stores schedule information for a sequence of statements (supplied
-    as a list of children).
-
-    :param children: the sequence of PSyIR nodes that make up the Schedule.
-    :type children: list of :py:class:`psyclone.psyGen.Node`
-    :param parent: that parent of this node in the PSyIR tree.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    '''
-    def __init__(self, children=None, parent=None):
-        Node.__init__(self, children=children, parent=parent)
-        self._text_name = "Schedule"
-        self._colour_key = "Schedule"
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: The name of this node in the dag.
-        :rtype: str
-        '''
-        return "schedule_" + str(self.abs_position)
-
-    def __getitem__(self, index):
-        '''
-        Overload the subscript notation ([int]) to access specific statements
-        in the Schedule.
-
-        :param int index: index of the statement to access.
-        :returns: statement in a given position in the Schedule sequence.
-        :rtype: :py:class:`psyclone.psyGen.Node`
-        '''
-        return self._children[index]
-
-    def __str__(self):
-        result = "Schedule:\n"
-        for entity in self._children:
-            result += str(entity) + "\n"
-        result += "End Schedule"
-        return result
-
-    def gen_code(self, parent):
-        '''
-        A Schedule does not have any direct Fortran representation. We just
-        call gen_code() for all of its children.
-
-        :param parent: node in the f2pygen AST to which to add content.
-        :type parent: :py:class:`psyclone.f2pygen.BaseGen`
-        '''
-        for child in self.children:
-            child.gen_code(parent)
 
 
 class InvokeSchedule(Schedule):
@@ -3285,493 +2396,6 @@ class HaloExchange(Node):
 
     def __str__(self):
         return self.node_str(False)
-
-
-class Loop(Node):
-    '''
-    Node representing a loop within the PSyIR. It has 4 mandatory children:
-    the first one represents the loop lower bound, the second one represents
-    the loop upper bound, the third one represents the step value and the
-    fourth one is always a PSyIR Schedule node containing the statements inside
-    the loop body.
-
-    (Note: currently this loop only represents the equivalent to Fortran do
-    loops. This means the loop is bounded by start/stop/step expressions
-    evaluated before the loop starts.)
-
-    :param parent: parent of this node in the PSyIR.
-    :type parent: sub-class of :py:class:`psyclone.psyGen.Node`
-    :param str variable_name: optional name of the loop iterator \
-        variable. Defaults to an empty string.
-    :param valid_loop_types: a list of loop types that are specific \
-        to a particular API.
-    :type valid_loop_types: list of str
-    :param annotations: One or more labels that provide additional information\
-          about the node (primarily relating to the input code that it was \
-          created from).
-    :type annotations: list of str
-
-    :raises InternalError: if the 'was_single_stmt' annotation is supplied \
-                           without the 'was_where' annotation.
-
-    '''
-    valid_annotations = ('was_where', 'was_single_stmt')
-
-    def __init__(self, parent=None, variable_name="", valid_loop_types=None,
-                 annotations=None):
-        Node.__init__(self, parent=parent, annotations=annotations)
-
-        # Although the base class checks on the annotations individually, we
-        # need to do further checks here
-        if annotations:
-            if 'was_single_stmt' in annotations and \
-               'was_where' not in annotations:
-                raise InternalError(
-                    "A Loop with the 'was_single_stmt' annotation "
-                    "must also have the 'was_where' annotation but"
-                    " got: {0}".format(annotations))
-
-        # we need to determine whether this is a built-in or kernel
-        # call so our schedule can do the right thing.
-
-        if valid_loop_types is None:
-            self._valid_loop_types = []
-        else:
-            self._valid_loop_types = valid_loop_types
-        self._loop_type = None        # inner, outer, colour, colours, ...
-        self._field = None
-        self._field_name = None       # name of the field
-        self._field_space = None      # v0, v1, ...,     cu, cv, ...
-        self._iteration_space = None  # cells, ...,      cu, cv, ...
-        self._kern = None             # Kernel associated with this loop
-        self._text_name = "Loop"
-        self._colour_key = "Loop"
-
-        # TODO replace iterates_over with iteration_space
-        self._iterates_over = "unknown"
-
-        self._variable_name = variable_name
-        self._id = ""
-
-    @staticmethod
-    def create(var_name, start, stop, step, children):
-        '''Create a Loop instance given valid instances of a variable name,
-        start, stop and step nodes, and a list of child nodes for the
-        loop body.
-
-        :param str var_name: the PSyIR node containing the variable \
-            name of the loop iterator.
-        :param start: the PSyIR node determining the value for the \
-            start of the loop.
-        :type start: :py:class:`psyclone.psyGen.Node`
-        :param end: the PSyIR node determining the value for the end \
-            of the loop.
-        :type end: :py:class:`psyclone.psyGen.Node`
-        :param step: the PSyIR node determining the value for the loop \
-            step.
-        :type step: :py:class:`psyclone.psyGen.Node`
-        :param children: a list of PSyIR nodes contained in the \
-            loop.
-        :type children: list of :py:class:`psyclone.psyGen.Node`
-
-        :returns: a Loop instance.
-        :rtype: :py:class:`psyclone.psyGen.Loop`
-
-        :raises GenerationError: if the arguments to the create method \
-            are not of the expected type.
-
-        '''
-        if not isinstance(var_name, str):
-            raise GenerationError(
-                "var_name argument in create method of Loop class "
-                "should be a string but found '{0}'."
-                "".format(type(var_name).__name__))
-        for name, instance in [("start", start), ("stop", stop),
-                               ("step", step)]:
-            if not isinstance(instance, Node):
-                raise GenerationError(
-                    "{0} argument in create method of Loop class should "
-                    "be a PSyIR Node but found '{1}'."
-                    "".format(name, type(instance).__name__))
-        if not isinstance(children, list):
-            raise GenerationError(
-                "children argument in create method of Loop class "
-                "should be a list but found '{0}'."
-                "".format(type(children).__name__))
-        for child in children:
-            if not isinstance(child, Node):
-                raise GenerationError(
-                    "child of children argument in create method of Loop "
-                    "class should be a PSyIR Node but found '{0}'."
-                    "".format(type(child).__name__))
-
-        loop = Loop()
-        start.parent = loop
-        stop.parent = loop
-        step.parent = loop
-        schedule = Schedule(parent=loop, children=children)
-        for child in children:
-            child.parent = schedule
-        loop.children = [start, stop, step, schedule]
-        loop._variable_name = var_name
-        return loop
-
-    def _check_completeness(self):
-        ''' Check that the Loop has 4 children and the 4th is a Schedule.
-
-        :raises InternalError: If the loop does not have 4 children or the
-            4th one is not a Schedule
-        '''
-        # We cannot just do str(self) in this routine we can end up being
-        # called as a result of str(self) higher up the call stack
-        # (because loop bounds are evaluated dynamically).
-        if len(self.children) < 4:
-            raise InternalError(
-                "Loop malformed or incomplete. It should have exactly 4 "
-                "children, but found loop with '{0}'.".format(
-                    ", ".join([str(child) for child in self.children])))
-
-        if not isinstance(self.children[3], Schedule):
-            raise InternalError(
-                "Loop malformed or incomplete. Fourth child should be a "
-                "Schedule node, but found loop with '{0}'.".format(
-                    ", ".join([str(child) for child in self.children])))
-
-    @property
-    def start_expr(self):
-        '''
-        :returns: the PSyIR Node representing the Loop start expression.
-        :rtype: :py:class:`psyclone.psyGen.Node`
-
-        '''
-        self._check_completeness()
-        return self._children[0]
-
-    @start_expr.setter
-    def start_expr(self, expr):
-        ''' Setter for Loop start_expr attribute.
-
-        :param expr: New PSyIR start expression.
-        :type expr: :py:class:`psyclone.psyGen.Node`
-
-        :raises TypeError: if expr is not a PSyIR node.
-
-        '''
-        if not isinstance(expr, Node):
-            raise TypeError(
-                "Only PSyIR nodes can be assigned as the Loop start expression"
-                ", but found '{0}' instead".format(type(expr)))
-        self._check_completeness()
-        self._children[0] = expr
-
-    @property
-    def stop_expr(self):
-        '''
-        :returns: the PSyIR Node representing the Loop stop expression.
-        :rtype: :py:class:`psyclone.psyGen.Node`
-
-        '''
-        self._check_completeness()
-        return self._children[1]
-
-    @stop_expr.setter
-    def stop_expr(self, expr):
-        ''' Setter for Loop stop_expr attribute.
-
-        :param expr: New PSyIR stop expression.
-        :type expr: :py:class:`psyclone.psyGen.Node`
-
-        :raises TypeError: if expr is not a PSyIR node.
-
-        '''
-        if not isinstance(expr, Node):
-            raise TypeError(
-                "Only PSyIR nodes can be assigned as the Loop stop expression"
-                ", but found '{0}' instead".format(type(expr)))
-        self._check_completeness()
-        self._children[1] = expr
-
-    @property
-    def step_expr(self):
-        '''
-        :returns: the PSyIR Node representing the Loop step expression.
-        :rtype: :py:class:`psyclone.psyGen.Node`
-
-        '''
-        self._check_completeness()
-        return self._children[2]
-
-    @step_expr.setter
-    def step_expr(self, expr):
-        ''' Setter for Loop step_expr attribute.
-
-        :param expr: New PSyIR step expression.
-        :type expr: :py:class:`psyclone.psyGen.Node`
-
-        :raises TypeError: if expr is not a PSyIR node.
-
-        '''
-        if not isinstance(expr, Node):
-            raise TypeError(
-                "Only PSyIR nodes can be assigned as the Loop step expression"
-                ", but found '{0}' instead".format(type(expr)))
-        self._check_completeness()
-        self._children[2] = expr
-
-    @property
-    def loop_body(self):
-        '''
-        :returns: the PSyIR Schedule with the loop body statements.
-        :rtype: :py:class:`psyclone.psyGen.Schedule`
-
-        '''
-        self._check_completeness()
-        return self._children[3]
-
-    @property
-    def dag_name(self):
-        ''' Return the name to use in a dag for this node
-
-        :returns: Return the dag name for this loop
-        :rtype: string
-
-        '''
-        if self.loop_type:
-            name = "loop_[{0}]_".format(self.loop_type) + \
-                   str(self.abs_position)
-        else:
-            name = "loop_" + str(self.abs_position)
-        return name
-
-    @property
-    def loop_type(self):
-        return self._loop_type
-
-    @loop_type.setter
-    def loop_type(self, value):
-        '''
-        Set the type of this Loop.
-
-        :param str value: the type of this loop.
-        :raises GenerationError: if the specified value is not a recognised \
-                                 loop type.
-        '''
-        if value not in self._valid_loop_types:
-            raise GenerationError(
-                "Error, loop_type value ({0}) is invalid. Must be one of "
-                "{1}.".format(value, self._valid_loop_types))
-        self._loop_type = value
-
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return ("{0}[type='{1}', field_space='{2}', it_space='{3}']".
-                format(colored("Loop", SCHEDULE_COLOUR_MAP["Loop"]),
-                       self._loop_type, self._field_space,
-                       self.iteration_space))
-
-    @property
-    def field_space(self):
-        return self._field_space
-
-    @field_space.setter
-    def field_space(self, my_field_space):
-        self._field_space = my_field_space
-
-    @property
-    def field_name(self):
-        return self._field_name
-
-    @property
-    def field(self):
-        return self._field
-
-    @field_name.setter
-    def field_name(self, my_field_name):
-        self._field_name = my_field_name
-
-    @property
-    def iteration_space(self):
-        return self._iteration_space
-
-    @iteration_space.setter
-    def iteration_space(self, it_space):
-        self._iteration_space = it_space
-
-    @property
-    def kernel(self):
-        '''
-        :returns: the kernel object associated with this Loop (if any).
-        :rtype: :py:class:`psyclone.psyGen.Kern`
-        '''
-        return self._kern
-
-    @kernel.setter
-    def kernel(self, kern):
-        '''
-        Setter for kernel object associated with this loop.
-
-        :param kern: a kernel object.
-        :type kern: :py:class:`psyclone.psyGen.Kern`
-        '''
-        self._kern = kern
-
-    @property
-    def variable_name(self):
-        '''
-        :returns: the name of the control variable for this loop.
-        :rtype: str
-        '''
-        return self._variable_name
-
-    def __str__(self):
-        # Give Loop sub-classes a specialised name
-        name = self.__class__.__name__
-        result = name + "["
-        result += "id:'" + self._id
-        result += "', variable:'" + self._variable_name
-        if self.loop_type:
-            result += "', loop_type:'" + self._loop_type
-        result += "']\n"
-        for entity in self._children:
-            result += str(entity) + "\n"
-        result += "End " + name
-        return result
-
-    def reference_accesses(self, var_accesses):
-        '''Get all variable access information. It combines the data from
-        the loop bounds (start, stop and step), as well as the loop body.
-        The loop variable is marked as 'READ+WRITE' and references in start,
-        stop and step are marked as 'READ'.
-
-        :param var_accesses: VariablesAccessInfo instance that stores the \
-            information about variable accesses.
-        :type var_accesses: \
-            :py:class:`psyclone.core.access_info.VariablesAccessInfo`
-        '''
-
-        # It is important to first add the WRITE access, since this way
-        # the dependency analysis for declaring openmp private variables
-        # will automatically declare the loop variables to be private
-        # (write access before read)
-        var_accesses.add_access(self.variable_name, AccessType.WRITE, self)
-        var_accesses.add_access(self.variable_name, AccessType.READ, self)
-
-        # Accesses of the start/stop/step expressions
-        self.start_expr.reference_accesses(var_accesses)
-        self.stop_expr.reference_accesses(var_accesses)
-        self.step_expr.reference_accesses(var_accesses)
-        var_accesses.next_location()
-
-        for child in self.loop_body.children:
-            child.reference_accesses(var_accesses)
-            var_accesses.next_location()
-
-    def has_inc_arg(self):
-        ''' Returns True if any of the Kernels called within this
-        loop have an argument with INC access. Returns False otherwise '''
-        for kern_call in self.coded_kernels():
-            for arg in kern_call.arguments.args:
-                if arg.access == AccessType.INC:
-                    return True
-        return False
-
-    def unique_modified_args(self, arg_type):
-        '''Return all unique arguments of the given type from kernels inside
-        this loop that are modified.
-
-        :param str arg_type: the type of kernel argument (e.g. field, \
-                             operator) to search for.
-        :returns: all unique arguments of the given type from kernels inside \
-            this loop that are modified.
-        :rtype: list of :py:class:`psyclone.psyGen.DynKernelArgument`
-        '''
-        arg_names = []
-        args = []
-        for call in self.kernels():
-            for arg in call.arguments.args:
-                if arg.type.lower() == arg_type:
-                    if arg.access != AccessType.READ:
-                        if arg.name not in arg_names:
-                            arg_names.append(arg.name)
-                            args.append(arg)
-        return args
-
-    def args_filter(self, arg_types=None, arg_accesses=None, unique=False):
-        '''Return all arguments of type arg_types and arg_accesses. If these
-        are not set then return all arguments. If unique is set to
-        True then only return uniquely named arguments'''
-        all_args = []
-        all_arg_names = []
-        for call in self.kernels():
-            call_args = args_filter(call.arguments.args, arg_types,
-                                    arg_accesses)
-            if unique:
-                for arg in call_args:
-                    if arg.name not in all_arg_names:
-                        all_args.append(arg)
-                        all_arg_names.append(arg.name)
-            else:
-                all_args.extend(call_args)
-        return all_args
-
-    def gen_code(self, parent):
-        '''
-        Generate the Fortran Loop and any associated code.
-
-        :param parent: the node in the f2pygen AST to which to add content.
-        :type parent: :py:class:`psyclone.f2pygen.SubroutineGen`
-
-        '''
-        def is_unit_literal(expr):
-            ''' Check if the given expression is equal to the literal '1'.
-
-            :param expr: a PSyIR expression.
-            :type expr: :py:class:`psyclone.psyGen.Node`
-
-            :returns: True if it is equal to the literal '1', false otherwise.
-            '''
-            return isinstance(expr, Literal) and expr.value == '1'
-
-        if not self.is_openmp_parallel():
-            calls = self.reductions()
-            zero_reduction_variables(calls, parent)
-
-        if self.root.opencl or (is_unit_literal(self.start_expr) and
-                                is_unit_literal(self.stop_expr)):
-            # no need for a loop
-            for child in self.loop_body:
-                child.gen_code(parent)
-        else:
-            from psyclone.psyir.backend.fortran import FortranWriter
-            from psyclone.f2pygen import DoGen, DeclGen
-            # start/stop/step_expr are generated with the FortranWriter
-            # backend, the rest of the loop with f2pygen.
-            fwriter = FortranWriter()
-            if is_unit_literal(self.step_expr):
-                step_str = None
-            else:
-                step_str = fwriter(self.step_expr)
-
-            do = DoGen(parent, self._variable_name,
-                       fwriter(self.start_expr),
-                       fwriter(self.stop_expr),
-                       step_str)
-            # need to add do loop before children as children may want to add
-            # info outside of do loop
-            parent.add(do)
-            for child in self.loop_body:
-                child.gen_code(do)
-            my_decl = DeclGen(parent, datatype="integer",
-                              entity_decls=[self._variable_name])
-            parent.add(my_decl)
 
 
 class Kern(Node):
@@ -5423,187 +4047,6 @@ class DummyTransformation(Transformation):
         return None, None
 
 
-class IfBlock(Node):
-    '''
-    Node representing an if-block within the PSyIR. It has two mandatory
-    children: the first one represents the if-condition and the second one
-    the if-body; and an optional third child representing the else-body.
-
-    '''
-    # The valid annotations for this If node:
-    # 'was_elseif' to tag nested ifs originally written with the 'else if'
-    # languague syntactic construct;
-    # 'was_single_stmt' to tag ifs with a 1-statement body which were
-    # originally written in a single line;
-    # 'was_case' to tag a conditional structure which was originally written
-    # with the Fortran 'case' or C 'switch' syntactic constructs;
-    # 'was_where' - a conditional structure originally implied by a Fortran
-    # WHERE construct.
-    valid_annotations = ('was_elseif', 'was_single_stmt', 'was_case',
-                         'was_where')
-
-    def __init__(self, parent=None, annotations=None):
-        super(IfBlock, self).__init__(parent=parent)
-        if annotations:
-            for annotation in annotations:
-                if annotation in IfBlock.valid_annotations:
-                    self._annotations.append(annotation)
-                else:
-                    raise InternalError(
-                        "IfBlock with unrecognized annotation '{0}', valid "
-                        "annotations are: {1}.".format(
-                            annotation, IfBlock.valid_annotations))
-        self._text_name = "If"
-        self._colour_key = "If"
-
-    @property
-    def condition(self):
-        ''' Return the PSyIR Node representing the conditional expression
-        of this IfBlock.
-
-        :returns: IfBlock conditional expression.
-        :rtype: :py:class:`psyclone.psyGen.Node`
-        :raises InternalError: If the IfBlock node does not have the correct \
-            number of children.
-        '''
-        if len(self.children) < 2:
-            raise InternalError(
-                "IfBlock malformed or incomplete. It should have at least 2 "
-                "children, but found {0}.".format(len(self.children)))
-        return self._children[0]
-
-    @property
-    def if_body(self):
-        ''' Return the Schedule executed when the IfBlock evaluates to True.
-
-        :returns: Schedule to be executed when IfBlock evaluates to True.
-        :rtype: :py:class:`psyclone.psyGen.Schedule`
-        :raises InternalError: If the IfBlock node does not have the correct \
-            number of children.
-        '''
-
-        if len(self.children) < 2:
-            raise InternalError(
-                "IfBlock malformed or incomplete. It should have at least 2 "
-                "children, but found {0}.".format(len(self.children)))
-
-        return self._children[1]
-
-    @property
-    def else_body(self):
-        ''' If available return the Schedule executed when the IfBlock
-        evaluates to False, otherwise return None.
-
-        :returns: Schedule to be executed when IfBlock evaluates \
-            to False, if it doesn't exist returns None.
-        :rtype: :py:class:`psyclone.psyGen.Schedule` or NoneType
-        '''
-        if len(self._children) == 3:
-            return self._children[2]
-        return None
-
-    @staticmethod
-    def create(if_condition, if_body, else_body=None):
-        '''Create an IfBlock instance given valid instances of an
-        if_condition, an if_body and an optional else_body.
-
-        :param if_condition: the PSyIR node containing the if \
-            condition of the if block.
-        :type if_condition: :py:class:`psyclone.psyGen.Node`
-        :param if_body: the PSyIR nodes representing the if body of \
-            the if block.
-        :type if_body: list of :py:class:`psyclone.psyGen.Node`
-        :param else_body: PSyIR nodes representing the else body of the \
-            if block of None if there is no else body (defaults to None).
-        :type else_body: list of :py:class:`psyclone.psyGen.Node` or \
-            NoneType
-
-        :returns: an IfBlock instance.
-        :rtype: :py:class:`psyclone.psyGen.IfBlock`
-
-        :raises GenerationError: if the arguments to the create method \
-            are not of the expected type.
-
-        '''
-        if not isinstance(if_condition, Node):
-            raise GenerationError(
-                "if_condition argument in create method of IfBlock class "
-                "should be a PSyIR Node but found '{0}'."
-                "".format(type(if_condition).__name__))
-        if not (isinstance(if_body, list) and all(isinstance(child, Node)
-                                                  for child in if_body)):
-            raise GenerationError(
-                "if_body argument in create method of IfBlock class should be "
-                "a list of PSyIR Nodes but it is either not a list or "
-                "one of the list's children is not a Node.")
-        if else_body and \
-           not (isinstance(if_body, list) and
-                all(isinstance(child, Node) for child in else_body)):
-            raise GenerationError(
-                "else_body argument in create method of IfBlock class should "
-                "be a list of PSyIR Nodes but it is either not a list or "
-                "one of the list's children is not a Node.")
-
-        if_stmt = IfBlock()
-        if_schedule = Schedule(parent=if_stmt)
-        for node in if_body:
-            node.parent = if_schedule
-        if_schedule.children = if_body
-        if else_body:
-            else_schedule = Schedule(parent=if_stmt)
-            for node in else_body:
-                node.parent = else_schedule
-            else_schedule.children = else_body
-            if_stmt.children = [if_condition, if_schedule, else_schedule]
-        else:
-            if_stmt.children = [if_condition, if_schedule]
-        if_condition.parent = if_stmt
-        return if_stmt
-
-    def node_str(self, colour=True):
-        ''' Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        text = self.coloured_name(colour) + "["
-        if self.annotations:
-            text += "annotations='" + ','.join(self.annotations) + "'"
-        text += "]"
-        return text
-
-    def __str__(self):
-        result = "If[]\n"
-        for entity in self._children:
-            result += str(entity)
-        return result
-
-    def reference_accesses(self, var_accesses):
-        '''Get all variable access information. It combines the data from
-        the condition, if-body and (if available) else-body. This could
-        later be extended to handle cases where a variable is only written
-        in one of the two branches.
-
-        :param var_accesses: VariablesAccessInfo instance that stores the \
-            information about variable accesses.
-        :type var_accesses: \
-            :py:class:`psyclone.core.access_info.VariablesAccessInfo`
-        '''
-
-        # The first child is the if condition - all variables are read-only
-        self.condition.reference_accesses(var_accesses)
-        var_accesses.next_location()
-        self.if_body.reference_accesses(var_accesses)
-        var_accesses.next_location()
-
-        if self.else_body:
-            self.else_body.reference_accesses(var_accesses)
-            var_accesses.next_location()
-
-
 class ACCKernelsDirective(ACCDirective):
     '''
     Class representing the !$ACC KERNELS directive in the PSyIR.
@@ -5826,970 +4269,6 @@ class KernelSchedule(Schedule):
             result += str(entity)
         result += "End KernelSchedule\n"
         return result
-
-
-class CodeBlock(Node):
-    '''Node representing some generic Fortran code that PSyclone does not
-    attempt to manipulate. As such it is a leaf in the PSyIR and therefore
-    has no children.
-
-    :param fp2_nodes: list of fparser2 AST nodes representing the Fortran \
-                      code constituting the code block.
-    :type fp2_nodes: list of :py:class:`fparser.two.utils.Base`
-    :param structure: argument indicating whether this code block is a \
-    statement or an expression.
-    :type structure: :py:class:`psyclone.psyGen.CodeBlock.Structure`
-    :param parent: the parent node of this code block in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    '''
-    class Structure(Enum):
-        '''
-        Enumeration that captures the structure of the code block which
-        may be required when processing.
-
-        '''
-        # The Code Block comprises one or more Fortran statements
-        # (which themselves may contain expressions).
-        STATEMENT = 1
-        # The Code Block comprises one or more Fortran expressions.
-        EXPRESSION = 2
-
-    def __init__(self, fp2_nodes, structure, parent=None):
-        super(CodeBlock, self).__init__(parent=parent)
-        # Store a list of the parser objects holding the code associated
-        # with this block. We make a copy of the contents of the list because
-        # the list itself is a temporary product of the process of converting
-        # from the fparser2 AST to the PSyIR.
-        self._fp2_nodes = fp2_nodes[:]
-        # Store references back into the fparser2 AST
-        if fp2_nodes:
-            self.ast = self._fp2_nodes[0]
-            self.ast_end = self._fp2_nodes[-1]
-        else:
-            self.ast = None
-            self.ast_end = None
-        # Store the structure of the code block.
-        self._structure = structure
-
-    @property
-    def structure(self):
-        '''
-        :returns: whether this code block is a statement or an expression.
-        :rtype: :py:class:`psyclone.psyGen.CodeBlock.Structure`
-
-        '''
-        return self._structure
-
-    @property
-    def get_ast_nodes(self):
-        '''
-        :returns: the list of nodes associated with this code block in \
-        the original AST.
-        :rtype: list of subclass of \
-        `:py:classfparser.two.Fortran2003.Base`
-
-        '''
-        return self._fp2_nodes
-
-    def node_str(self, colour=True):
-        ''' Create a text description of this node in the schedule, optionally
-        including control codes for colour.
-
-        :param bool colour: whether or not to include control codes for colour.
-
-        :return: text description of this node.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + \
-            "[" + str(list(map(type, self._fp2_nodes))) + "]"
-
-    def __str__(self):
-        return "CodeBlock[{0} nodes]".format(len(self._fp2_nodes))
-
-
-class Assignment(Node):
-    '''
-    Node representing an Assignment statement. As such it has a LHS and RHS
-    as children 0 and 1 respectively.
-
-    :param ast: node in the fparser2 AST representing the assignment.
-    :type ast: :py:class:`fparser.two.Fortran2003.Assignment_Stmt.
-    :param parent: the parent node of this Assignment in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-    '''
-    def __init__(self, ast=None, parent=None):
-        super(Assignment, self).__init__(ast=ast, parent=parent)
-
-    @property
-    def lhs(self):
-        '''
-        :returns: the child node representing the Left-Hand Side of the \
-            assignment.
-        :rtype: :py:class:`psyclone.psyGen.Node`
-
-        :raises InternalError: Node has fewer children than expected.
-        '''
-        if not self._children:
-            raise InternalError(
-                "Assignment '{0}' malformed or incomplete. It "
-                "needs at least 1 child to have a lhs.".format(repr(self)))
-
-        return self._children[0]
-
-    @property
-    def rhs(self):
-        '''
-        :returns: the child node representing the Right-Hand Side of the \
-            assignment.
-        :rtype: :py:class:`psyclone.psyGen.Node`
-
-        :raises InternalError: Node has fewer children than expected.
-        '''
-        if len(self._children) < 2:
-            raise InternalError(
-                "Assignment '{0}' malformed or incomplete. It "
-                "needs at least 2 children to have a rhs.".format(repr(self)))
-
-        return self._children[1]
-
-    @staticmethod
-    def create(lhs, rhs):
-        '''Create an Assignment instance given lhs and rhs child instances.
-
-        :param lhs: the PSyIR node containing the left hand side of \
-            the assignment.
-        :type lhs: :py:class:`psyclone.psyGen.Node`
-        :param rhs: the PSyIR node containing the right hand side of \
-            the assignment.
-        :type rhs: :py:class:`psyclone.psyGen.Node`
-
-        :returns: an Assignment instance.
-        :rtype: :py:class:`psyclone.psyGen.Assignment`
-
-        :raises GenerationError: if the arguments to the create method \
-            are not of the expected type.
-
-        '''
-        for name, instance in [("lhs", lhs), ("rhs", rhs)]:
-            if not isinstance(instance, Node):
-                raise GenerationError(
-                    "{0} argument in create method of Assignment class "
-                    "should be a PSyIR Node but found '{1}'."
-                    "".format(name, type(instance).__name__))
-
-        new_assignment = Assignment()
-        lhs.parent = new_assignment
-        rhs.parent = new_assignment
-        new_assignment.children = [lhs, rhs]
-        return new_assignment
-
-    def __str__(self):
-        result = "Assignment[]\n"
-        for entity in self._children:
-            result += str(entity)
-        return result
-
-    def reference_accesses(self, var_accesses):
-        '''Get all variable access information from this node. The assigned-to
-        variable will be set to 'WRITE'.
-
-        :param var_accesses: VariablesAccessInfo instance that stores the \
-            information about variable accesses.
-        :type var_accesses: \
-            :py:class:`psyclone.core.access_info.VariablesAccessInfo`
-        '''
-
-        # It is important that a new instance is used to handle the LHS,
-        # since a check in 'change_read_to_write' makes sure that there
-        # is only one access to the variable!
-        accesses_left = VariablesAccessInfo()
-        self.lhs.reference_accesses(accesses_left)
-
-        # Now change the (one) access to the assigned variable to be WRITE:
-        if isinstance(self.lhs, CodeBlock):
-            # TODO #363: Assignment to user defined type, not supported yet.
-            # Here an absolute hack to get at least some information out
-            # from the AST - though indices are just strings, which will
-            # likely cause problems later as well.
-            name = str(self.lhs.ast)
-            # A regular expression that tries to find the last parenthesis
-            # pair in the name ("a(i,j)" --> "(i,j)")
-            ind = re.search(r"\([^\(]+\)$", name)
-            if ind:
-                # Remove the index part of the name
-                name = name.replace(ind.group(0), "")
-                # The index must be added as a list
-                accesses_left.add_access(name, AccessType.WRITE, self,
-                                         [ind.group(0)])
-            else:
-                accesses_left.add_access(name, AccessType.WRITE, self)
-        else:
-            var_info = accesses_left[self.lhs.name]
-            try:
-                var_info.change_read_to_write()
-            except InternalError:
-                # An internal error typically indicates that the same variable
-                # is used twice on the LHS, e.g.: g(g(1)) = ... This is not
-                # supported in PSyclone.
-                from psyclone.parse.utils import ParseError
-                raise ParseError("The variable '{0}' appears more than once "
-                                 "on the left-hand side of an assignment."
-                                 .format(self.lhs.name))
-
-        # Merge the data (that shows now WRITE for the variable) with the
-        # parameter to this function. It is important that first the
-        # RHS is added, so that in statements like 'a=a+1' the read on
-        # the RHS comes before the write on the LHS (they have the same
-        # location otherwise, but the order is still important)
-        self.rhs.reference_accesses(var_accesses)
-        var_accesses.merge(accesses_left)
-        var_accesses.next_location()
-
-
-class Reference(Node):
-    '''
-    Node representing a Reference Expression.
-
-    :param str reference_name: the name of the symbol being referenced.
-    :param parent: the parent node of this Reference in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node` or NoneType
-    '''
-    def __init__(self, reference_name, parent=None):
-        super(Reference, self).__init__(parent=parent)
-        self._reference = reference_name
-
-    @property
-    def name(self):
-        ''' Return the name of the referenced symbol.
-
-        :returns: Name of the referenced symbol.
-        :rtype: str
-        '''
-        return self._reference
-
-    def node_str(self, colour=True):
-        ''' Create a text description of this node in the schedule, optionally
-        including control codes for colour.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :return: text description of this node.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[name:'" + self._reference + "']"
-
-    def __str__(self):
-        return self.node_str(False)
-
-    def math_equal(self, other):
-        ''':param other: the node to compare self with.
-        :type other: py:class:`psyclone.psyGen.Node`
-
-        :returns: True if the self has the same results as other.
-        :rtype: bool
-        '''
-        if not super(Reference, self).math_equal(other):
-            return False
-        return self.name == other.name
-
-    def reference_accesses(self, var_accesses):
-        '''Get all variable access information from this node, i.e.
-        it sets this variable to be read.
-
-        :param var_accesses: VariablesAccessInfo instance that stores the \
-            information about variable accesses.
-        :type var_accesses: \
-            :py:class:`psyclone.core.access_info.VariablesAccessInfo`
-        '''
-        var_accesses.add_access(self._reference, AccessType.READ, self)
-
-    def check_declared(self):
-        '''Check whether this reference has an associated symbol table entry.
-
-        raises SymbolError: if one or more ancestor symbol table(s) \
-        are found and the name of this reference is not found in any \
-        of them.
-
-        '''
-        found_symbol_table = False
-        test_node = self.parent
-        while test_node:
-            if hasattr(test_node, 'symbol_table'):
-                found_symbol_table = True
-                symbol_table = test_node.symbol_table
-                if self.name in symbol_table:
-                    return
-            test_node = test_node.parent
-
-        # TODO: remove this if test, remove the initialisation of the
-        # found_symbol_table boolean variable and update the doc
-        # string when SymbolTables are suppported in the NEMO API, see
-        # issue #500. After this change has been made this method could
-        # make use of the symbol method to determine
-        # whether the reference has been declared (or not).
-        if found_symbol_table:
-            raise SymbolError(
-                "Undeclared reference '{0}' found.".format(self.name))
-
-    def symbol(self, scope_limit=None):
-        '''Returns the symbol from a symbol table associated with this
-        reference or None is it is not found. The scope_limit variable
-        limits the symbol table search to nodes within the scope.
-
-        :param scope_limit: optional Node which limits the symbol \
-        search space to the symbol tables of the nodes within the \
-        given scope. If it is None (the default), the whole scope (all \
-        symbol tables in ancestor nodes) is searched.
-        :type scope_limit: :py:class:`psyclone.psyGen.Node` or `None`
-
-        :returns: the Symbol associated with this reference if one is \
-        found or None if not.
-        :rtype: :py:class:`psyclone.psyir.symbols.Symbol` or `None`
-
-        '''
-        if scope_limit:
-
-            if not isinstance(scope_limit, Node):
-                raise TypeError(
-                    "The scope_limit argument '{0}' provided to the symbol "
-                    "method, is not of type `Node`."
-                    "".format(str(scope_limit), str(self)))
-
-            # Check that the scope_limit Node is an ancestor of this
-            # Reference Node and raise an exception if not.
-            found = False
-            mynode = self.parent
-            while mynode is not None:
-                if mynode is scope_limit:
-                    found = True
-                    break
-                mynode = mynode.parent
-            if not found:
-                # The scope_limit node is not an ancestor of this reference
-                # so raise an exception.
-                raise ValueError(
-                    "The scope_limit node '{0}' provided to the symbol "
-                    "method, is not an ancestor of this reference node '{1}'."
-                    "".format(str(scope_limit), str(self)))
-        test_node = self.parent
-        # Iterate over ancestor Nodes of this Reference Node.
-        while test_node:
-            # For simplicity, test every Node for the existence of a
-            # SymbolTable (rather than checking for the particular
-            # Node types which we know to have SymbolTables).
-            if hasattr(test_node, 'symbol_table'):
-                # This Node does have a SymbolTable.
-                symbol_table = test_node.symbol_table
-                try:
-                    # If the reference matches a Symbol in this
-                    # SymbolTable then return the Symbol.
-                    return symbol_table.lookup(self.name)
-                except KeyError:
-                    # The Reference Node does not match any Symbols in
-                    # this SymbolTable.
-                    pass
-            if test_node is scope_limit:
-                # The ancestor scope Node has been reached and nothing
-                # has matched so return with None.
-                return None
-            # Move on to the next ancestor.
-            test_node = test_node.parent
-        # scope has not been set and all Nodes have been checked (up
-        # to the root Node) but there has been no match so return with
-        # None.
-        return None
-
-
-@six.add_metaclass(abc.ABCMeta)
-class Operation(Node):
-    '''
-    Abstract base class for PSyIR nodes representing operators.
-
-    :param operator: the operator used in the operation.
-    :type operator: :py:class:`psyclone.psyGen.UnaryOperation.Operator` or \
-                    :py:class:`psyclone.psyGen.BinaryOperation.Operator` or \
-                    :py:class:`psyclone.psyGen.NaryOperation.Operator`
-    :param parent: the parent node of this Operation in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    :raises TypeError: if the supplied operator is not an instance of \
-                       self.Operator.
-
-    '''
-    # Must be overridden in sub-class to hold an Enumeration of the Operators
-    # that it can represent.
-    Operator = None
-
-    def __init__(self, operator, parent=None):
-        super(Operation, self).__init__(parent=parent)
-
-        if not isinstance(operator, self.Operator):
-            raise TypeError(
-                "{0} operator argument must be of type "
-                "{0}.Operator but found {1}.".format(type(self).__name__,
-                                                     type(operator).__name__))
-        self._operator = operator
-        self._text_name = "Operation"
-        self._colour_key = "Operation"
-
-    @property
-    def operator(self):
-        '''
-        Return the operator.
-
-        :returns: Enumerated type capturing the operator.
-        :rtype: :py:class:`psyclone.psyGen.UnaryOperation.Operator` or \
-                :py:class:`psyclone.psyGen.BinaryOperation.Operator` or \
-                :py:class:`psyclone.psyGen.NaryOperation.Operator`
-
-        '''
-        return self._operator
-
-    def node_str(self, colour=True):
-        '''
-        Construct a text representation of this node, optionally with control
-        codes for coloured display in a suitable terminal.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this PSyIR node.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + \
-            "[operator:'" + self._operator.name + "']"
-
-    def __str__(self):
-        result = self.node_str(False) + "\n"
-        for entity in self._children:
-            result += str(entity) + "\n"
-
-        # Delete last line break
-        if result[-1] == "\n":
-            result = result[:-1]
-        return result
-
-
-class UnaryOperation(Operation):
-    '''
-    Node representing a UnaryOperation expression. As such it has one operand
-    as child 0, and an attribute with the operator type.
-
-    :param operator: Enumerated type capturing the unary operator.
-    :type operator: :py:class:`psyclone.psyGen.UnaryOperation.Operator`
-    :param parent: the parent node of this UnaryOperation in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    '''
-    Operator = Enum('Operator', [
-        # Arithmetic Operators
-        'MINUS', 'PLUS', 'SQRT', 'EXP', 'LOG', 'LOG10',
-        # Logical Operators
-        'NOT',
-        # Trigonometric Operators
-        'COS', 'SIN', 'TAN', 'ACOS', 'ASIN', 'ATAN',
-        # Other Maths Operators
-        'ABS', 'CEIL',
-        # Casting Operators
-        'REAL', 'INT'
-        ])
-
-    def __init__(self, operation, parent=None):
-        super(UnaryOperation, self).__init__(operation, parent)
-        self._text_name = "UnaryOperation"
-
-    @staticmethod
-    def create(oper, child):
-        '''Create a UnaryOperation instance given oper and child instances.
-
-        :param oper: the specified operator.
-        :type oper: :py:class:`psyclone.psyGen.UnaryOperation.Operator`
-        :param child: the PSyIR node that oper operates on.
-        :type child: :py:class:`psyclone.psyGen.Node`
-
-        :returns: a UnaryOperation instance.
-        :rtype: :py:class:`psyclone.psyGen.UnaryOperation`
-
-        :raises GenerationError: if the arguments to the create method \
-            are not of the expected type.
-
-        '''
-        if not isinstance(oper, UnaryOperation.Operator):
-            raise GenerationError(
-                "oper argument in create method of UnaryOperation class "
-                "should be a PSyIR UnaryOperation Operator but found '{0}'."
-                "".format(type(oper).__name__))
-        if not isinstance(child, Node):
-            raise GenerationError(
-                "child argument in create method of UnaryOperation class "
-                "should be a PSyIR Node but found '{0}'."
-                "".format(type(child).__name__))
-
-        unary_op = UnaryOperation(oper)
-        child.parent = unary_op
-        unary_op.children = [child]
-        return unary_op
-
-
-class BinaryOperation(Operation):
-    '''
-    Node representing a BinaryOperation expression. As such it has two operands
-    as children 0 and 1, and an attribute with the operator type.
-
-    :param operator: the operator used in the operation.
-    :type operator: :py:class:`psyclone.psyGen.BinaryOperation.Operator`
-    :param parent: the parent node of this Operation in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    '''
-    Operator = Enum('Operator', [
-        # Arithmetic Operators. ('REM' is remainder AKA 'MOD' in Fortran.)
-        'ADD', 'SUB', 'MUL', 'DIV', 'REM', 'POW', 'SUM',
-        # Relational Operators
-        'EQ', 'NE', 'GT', 'LT', 'GE', 'LE',
-        # Logical Operators
-        'AND', 'OR',
-        # Other Maths Operators
-        'SIGN', 'MIN', 'MAX',
-        # Query Operators
-        'SIZE'
-        ])
-
-    def __init__(self, operator, parent=None):
-        super(BinaryOperation, self).__init__(operator, parent)
-        self._text_name = "BinaryOperation"
-
-    def math_equal(self, other):
-        ''':param other: the node to compare self with.
-        :type other: py:class:`psyclone.psyGen.Node`
-        :returns: True if the self has the same results as other.
-        :rtype: bool
-        '''
-        if not super(BinaryOperation, self).math_equal(other):
-            # Support some commutative law, unfortunately we now need
-            # to repeat some tests already done in super(), since we
-            # don't know why the above test failed
-            # TODO #533 for documenting restrictions
-            # pylint: disable=unidiomatic-typecheck
-            if type(self) != type(other):
-                return False
-            if self.operator != other.operator:
-                return False
-            if self.operator not in [self.Operator.ADD, self.Operator.MUL,
-                                     self.Operator.AND, self.Operator.OR,
-                                     self.Operator.EQ]:
-                return False
-            return self._children[0].math_equal(other.children[1]) and \
-                self._children[1].math_equal(other.children[0])
-        return self.operator == other.operator
-
-    @staticmethod
-    def create(oper, lhs, rhs):
-        '''Create a BinaryOperator instance given an operator and lhs and rhs
-        child instances.
-
-        :param operator: the operator used in the operation.
-        :type operator: :py:class:`psyclone.psyGen.BinaryOperation.Operator`
-        :param lhs: the PSyIR node containing the left hand side of \
-            the assignment.
-        :type lhs: :py:class:`psyclone.psyGen.Node`
-        :param rhs: the PSyIR node containing the right hand side of \
-            the assignment.
-        :type rhs: :py:class:`psyclone.psyGen.Node`
-
-        :returns: a BinaryOperator instance.
-        :rtype: :py:class:`psyclone.psyGen.BinaryOperator`
-
-        :raises GenerationError: if the arguments to the create method \
-            are not of the expected type.
-
-        '''
-        if not isinstance(oper, BinaryOperation.Operator):
-            raise GenerationError(
-                "oper argument in create method of BinaryOperation class "
-                "should be a PSyIR BinaryOperation Operator but found '{0}'."
-                "".format(type(oper).__name__))
-        for name, instance in [("lhs", lhs), ("rhs", rhs)]:
-            if not isinstance(instance, Node):
-                raise GenerationError(
-                    "{0} argument in create method of BinaryOperation class "
-                    "should be a PSyIR Node but found '{1}'."
-                    "".format(name, type(instance).__name__))
-
-        binary_op = BinaryOperation(oper)
-        lhs.parent = binary_op
-        rhs.parent = binary_op
-        binary_op.children = [lhs, rhs]
-        return binary_op
-
-
-class NaryOperation(Operation):
-    '''
-    Node representing a n-ary operation expression. The n operands are the
-    stored as the 0 - n-1th children of this node and the type of the operator
-    is held in an attribute.
-
-
-    :param operator: the operator used in the operation.
-    :type operator: :py:class:`psyclone.psyGen.NaryOperation.Operator`
-    :param parent: the parent node of this Operation in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    '''
-    Operator = Enum('Operator', [
-        # Arithmetic Operators
-        'MAX', 'MIN', 'SUM'
-        ])
-
-    def __init__(self, operator, parent=None):
-        super(NaryOperation, self).__init__(operator, parent)
-        self._text_name = "NaryOperation"
-
-    @staticmethod
-    def create(oper, children):
-        '''Create an NaryOperator instance given an operator and a list of
-        Node instances.
-
-        :param operator: the operator used in the operation.
-        :type operator: :py:class:`psyclone.psyGen.NaryOperation.Operator`
-        :param children: a list of PSyIR nodes that the operator \
-            operates on.
-        :type children: list of :py:class:`psyclone.psyGen.Node`
-
-        :returns: an NaryOperator instance.
-        :rtype: :py:class:`psyclone.psyGen.NaryOperator`
-
-        :raises GenerationError: if the arguments to the create method \
-            are not of the expected type.
-
-        '''
-        if not isinstance(oper, NaryOperation.Operator):
-            raise GenerationError(
-                "oper argument in create method of NaryOperation class "
-                "should be a PSyIR NaryOperation Operator but found '{0}'."
-                "".format(type(oper).__name__))
-        if not isinstance(children, list):
-            raise GenerationError(
-                "children argument in create method of NaryOperation class "
-                "should be a list but found '{0}'."
-                "".format(type(children).__name__))
-        for child in children:
-            if not isinstance(child, Node):
-                raise GenerationError(
-                    "child of children argument in create method of "
-                    "NaryOperation class should be a PSyIR Node but "
-                    "found '{0}'.".format(type(child).__name__))
-
-        nary_op = NaryOperation(oper)
-        for child in children:
-            child.parent = nary_op
-        nary_op.children = children
-        return nary_op
-
-
-class Array(Reference):
-    '''
-    Node representing an Array reference. As such it has a reference and a
-    subscript list as children 0 and 1, respectively.
-
-    :param str reference_name: name of the array symbol.
-    :param parent: the parent node of this Array in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    '''
-    def __init__(self, reference_name, parent=None):
-        super(Array, self).__init__(reference_name, parent=parent)
-        self._text_name = "ArrayReference"
-        self._colour_key = "Reference"
-
-    @staticmethod
-    def create(name, children):
-        '''Create an Array instance given an array name and a list of Node
-        array indices.
-
-        :param str name: the name of the array.
-        :param children: a list of Nodes describing the array indices.
-        :type children: list of :py:class:`psyclone.psyGen.Node`
-
-        :returns: an Array instance.
-        :rtype: :py:class:`psyclone.psyGen.Array`
-
-        :raises GenerationError: if the arguments to the create method \
-            are not of the expected type.
-
-        '''
-        if not isinstance(name, str):
-            raise GenerationError(
-                "name argument in create method of Array class "
-                "should be a string but found '{0}'."
-                "".format(type(name).__name__))
-        if not name:
-            raise GenerationError(
-                "name argument in create method of Array class "
-                "can't be an empty string.")
-        if not isinstance(children, list):
-            raise GenerationError(
-                "children argument in create method of Array class "
-                "should be a list but found '{0}'."
-                "".format(type(children).__name__))
-        for child in children:
-            if not isinstance(child, Node):
-                raise GenerationError(
-                    "child of children argument in create method of "
-                    "Array class should be a PSyIR Node but found '{0}'."
-                    "".format(type(child).__name__))
-
-        array = Array(name)
-        for child in children:
-            child.parent = array
-        array.children = children
-        return array
-
-    def __str__(self):
-        result = "Array" + super(Array, self).__str__() + "\n"
-        for entity in self._children:
-            result += str(entity) + "\n"
-        return result
-
-    def reference_accesses(self, var_accesses):
-        '''Get all variable access information. All variables used as indices
-        in the access of the array will be added as READ.
-        :param var_accesses: \
-            :py:class:`psyclone.core.access_info.VariablesAccessInfo`
-        '''
-
-        # This will set the array-name as READ
-        super(Array, self).reference_accesses(var_accesses)
-
-        # Now add all children: Note that the class Reference
-        # does not recurse to the children (which store the indices), so at
-        # this stage no index information has been stored:
-        list_indices = []
-        for child in self._children:
-            child.reference_accesses(var_accesses)
-            list_indices.append(child)
-
-        if list_indices:
-            var_info = var_accesses[self._reference]
-            # The last entry in all_accesses is the one added above
-            # in super(Array...). Add the indices to that entry.
-            var_info.all_accesses[-1].indices = list_indices
-
-
-class Literal(Node):
-    '''
-    Node representing a Literal. The value and datatype properties of
-    this node are immutable.
-
-    :param str value: the value of the literal.
-    :param datatype: the datatype of this literal.
-    :type datatype: :py:class:`psyclone.psyir.symbols.DataType`
-    :param parent: the parent node of this Literal in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    :raises TypeError: if the datatype is not an instance of \
-                       :py:class:`psyclone.psyir.symbols.DataType`.
-    :raises ValueError: if the datatype is not one of \
-                        self.VALID_DATA_TYPES.
-    :raises TypeError: if the supplied value is not a string.
-    :raises ValueError: if the Literal is a BOOLEAN and the value is not \
-                        'true' or 'false'.
-    '''
-    # A Literal cannot have DEFERRED type
-    VALID_DATA_TYPES = [DataType.INTEGER, DataType.REAL,
-                        DataType.CHARACTER, DataType.BOOLEAN]
-
-    def __init__(self, value, datatype, parent=None):
-        super(Literal, self).__init__(parent=parent)
-
-        # Checks for the datatype
-        if not isinstance(datatype, DataType):
-            raise TypeError("The datatype of a Literal must be an instance of"
-                            " psyir.symbols.DataType but got '{0}'".format(
-                                type(datatype).__name__))
-        if datatype not in self.VALID_DATA_TYPES:
-            raise ValueError("The datatype of a Literal must be one of {0} "
-                             "but got '{1}'".format(self.VALID_DATA_TYPES,
-                                                    datatype))
-        if not isinstance(value, six.string_types):
-            raise TypeError("Literals must be supplied with "
-                            "a value encoded as a string but got: {0}".
-                            format(type(value).__name__))
-
-        if datatype is DataType.BOOLEAN:
-            if value not in ("true", "false"):
-                raise ValueError(
-                    "A DataType.BOOLEAN Literal can only be: 'true' or "
-                    "'false' but got '{0}' instead.".format(value))
-
-        self._datatype = datatype
-        self._value = value
-
-    @property
-    def datatype(self):
-        '''
-        :returns: the type of this Literal.
-        :rtype: :py:class:`psyclone.psyGen.DataType`
-        '''
-        return self._datatype
-
-    @property
-    def value(self):
-        '''
-        :returns: String representing the literal value.
-        :rtype: str
-        '''
-        return self._value
-
-    def node_str(self, colour=True):
-        '''
-        Construct a text representation of this node, optionally containing
-        colour control codes.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this PSyIR node.
-        :rtype: str
-        '''
-        return "{0}[value:'{1}', {2}]".format(
-            self.coloured_name(colour),
-            self._value, str(self.datatype))
-
-    def math_equal(self, other):
-        ''':param other: the node to compare self with.
-        :type other: py:class:`psyclone.psyGen.Node`
-
-        :return: if the self has the same results as other.
-        :type: bool
-        '''
-
-        return self.value == other.value
-
-
-class Return(Node):
-    '''
-    Node representing a Return statement (subroutine break without return
-    value).
-
-    :param parent: the parent node of this Return in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-    '''
-    def __init__(self, parent=None):
-        super(Return, self).__init__(parent=parent)
-
-    def __str__(self):
-        return "Return[]\n"
-
-
-class Container(Node):
-    '''Node representing a set of KernelSchedule and/or Container nodes,
-    as well as a name and a SymbolTable. This construct can be used to
-    scope symbols of variables, KernelSchedule names and Container
-    names. In Fortran a container would naturally represent a module
-    or a submodule.
-
-    :param str name: the name of the container.
-    :param parent: optional parent node of this Container in the PSyIR.
-    :type parent: :py:class:`psyclone.psyGen.Node`
-
-    '''
-    def __init__(self, name, parent=None):
-        super(Container, self).__init__(parent=parent)
-        self._name = name
-        self._symbol_table = SymbolTable(self)
-
-    @staticmethod
-    def create(name, symbol_table, children):
-        '''Create a Container instance given a name, a symbol table and a
-        list of child nodes.
-
-        :param str name: the name of the Container.
-        :param symbol_table: the symbol table associated with this \
-            Container.
-        :type symbol_table: :py:class:`psyclone.psyGen.SymbolTable`
-        :param children: a list of PSyIR nodes contained in the \
-            Container. These must be Containers or KernelSchedules.
-        :type children: list of :py:class:`psyclone.psyGen.Container` \
-            or :py:class:`psyclone.psyGen.KernelSchedule`
-
-        :returns: a Container instance.
-        :rtype: :py:class:`psyclone.psyGen.Container`
-
-        :raises GenerationError: if the arguments to the create method \
-            are not of the expected type.
-
-        '''
-        if not isinstance(name, str):
-            raise GenerationError(
-                "name argument in create method of Container class "
-                "should be a string but found '{0}'."
-                "".format(type(name).__name__))
-        if not isinstance(symbol_table, SymbolTable):
-            raise GenerationError(
-                "symbol_table argument in create method of Container class "
-                "should be a SymbolTable but found '{0}'."
-                "".format(type(symbol_table).__name__))
-        if not isinstance(children, list):
-            raise GenerationError(
-                "children argument in create method of Container class "
-                "should be a list but found '{0}'."
-                "".format(type(children).__name__))
-        for child in children:
-            if not isinstance(child, (KernelSchedule, Container)):
-                raise GenerationError(
-                    "child of children argument in create method of "
-                    "Container class should be a PSyIR KernelSchedule or "
-                    "Container but found '{0}'.".format(type(child).__name__))
-
-        container = Container(name)
-        container._symbol_table = symbol_table
-        symbol_table._schedule = container
-        for child in children:
-            child.parent = container
-        container.children = children
-        return container
-
-    @property
-    def name(self):
-        '''
-        :returns: name of the container.
-        :rtype: str
-
-        '''
-        return self._name
-
-    @name.setter
-    def name(self, new_name):
-        '''Sets a new name for the container.
-
-        :param str new_name: new name for the container.
-
-        '''
-        self._name = new_name
-
-    @property
-    def symbol_table(self):
-        '''
-        :returns: table containing symbol information for the container.
-        :rtype: :py:class:`psyclone.psyir.symbols.SymbolTable`
-
-        '''
-        return self._symbol_table
-
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with appropriate control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[{0}]".format(self.name)
-
-    def __str__(self):
-        return "Container[{0}]\n".format(self.name)
 
 
 __all__ = ['UnaryOperation', 'BinaryOperation', 'NaryOperation',
