+++ src/psyclone/psyir/backend/sir.py
@@ -41,7 +41,7 @@ gen() method to generate Fortran.
 '''
 
 from psyclone.psyir.backend.visitor import PSyIRVisitor, VisitorError
-from psyclone.psyGen import Reference, BinaryOperation, Literal, \
+from psyclone.psyir.nodes import Reference, BinaryOperation, Literal, \
     Array, UnaryOperation
 from psyclone.nemo import NemoLoop, NemoKern
 from psyclone.psyir.symbols import DataType
