+++ src/psyclone/tests/domain/gocean/transformations/gocean1p0_transformations_test.py
@@ -42,7 +42,7 @@ import os
 import re
 import pytest
 from psyclone.configuration import Config
-from psyclone.psyGen import Loop
+from psyclone.psyir.nodes import Loop
 from psyclone.psyir.transformations import TransformationError
 from psyclone.transformations import GOConstLoopBoundsTrans, \
     LoopFuseTrans, GOLoopSwapTrans, OMPParallelTrans, \
