+++ src/psyclone/tests/psyir/backend/fortran_test.py
@@ -44,7 +44,7 @@ from fparser.common.readfortran import FortranStringReader
 from psyclone.psyir.backend.visitor import VisitorError
 from psyclone.psyir.backend.fortran import gen_intent, gen_dims, \
     FortranWriter, gen_datatype
-from psyclone.psyGen import Node, CodeBlock, Container, Literal
+from psyclone.psyir.nodes import Node, CodeBlock, Container, Literal
 from psyclone.psyir.symbols import DataSymbol, SymbolTable, ContainerSymbol, \
     GlobalInterface, ArgumentInterface, UnresolvedInterface, DataType
 from psyclone.tests.utilities import create_schedule
