+++ src/psyclone/tests/psyir/backend/psyir_openmp_test.py
@@ -38,7 +38,7 @@
 
 from __future__ import absolute_import
 
-from psyclone.psyGen import Assignment, Reference
+from psyclone.psyir.nodes import Assignment, Reference
 from psyclone.psyir.backend.c import CWriter
 from psyclone.psyir.backend.fortran import FortranWriter
 from psyclone.tests.utilities import create_schedule, get_invoke
