+++ src/psyclone/tests/psyir/backend/sir_test.py
@@ -123,7 +123,7 @@ def get_assignment(parser, code):
     :rtype: :py:class:`psyclone.psyGen.Assignment`
 
     '''
-    from psyclone.psyGen import Assignment
+    from psyclone.psyir.nodes import Assignment
     kernel = get_kernel(parser, code)
     kernel_schedule = kernel.get_kernel_schedule()
     assignment = kernel_schedule.children[0]
@@ -253,7 +253,7 @@ def test_sirwriter_node_1(parser):
     True. Also check for SIR indentation.
 
     '''
-    from psyclone.psyGen import Node
+    from psyclone.psyir.nodes import Node
     schedule = get_schedule(parser, CODE)
 
     # pylint: disable=abstract-method
@@ -904,7 +904,7 @@ def test_sirwriter_schedule_node_1(parser, sir_writer):
     creates the expected code by calling its children.
 
     '''
-    from psyclone.psyGen import Schedule
+    from psyclone.psyir.nodes import Schedule
     code = CODE.replace(
         "a(i,j,k) = 1.0", "if (b .eq. c) then\na(i,j,k) = 1.0\nend if")
     kernel = get_kernel(parser, code)
