+++ src/psyclone/tests/psyir/frontend/literals_test.py
@@ -43,7 +43,7 @@ from fparser.common.readfortran import FortranStringReader
 from fparser.two import Fortran2003
 from psyclone.psyir.frontend.fparser2 import Fparser2Reader
 from psyclone.psyir.symbols import DataType
-from psyclone.psyGen import Node, Literal, CodeBlock
+from psyclone.psyir.nodes import Node, Literal, CodeBlock
 
 
 @pytest.mark.parametrize("code, dtype", [("'hello'", DataType.CHARACTER),
