+++ src/psyclone/tests/psyir/transformations/profile_test.py
@@ -44,7 +44,8 @@ import pytest
 
 from psyclone.generator import GenerationError
 from psyclone.profiler import Profiler, ProfileNode
-from psyclone.psyGen import InternalError, Loop, NameSpace
+from psyclone.psyir.nodes import Loop
+from psyclone.psyGen import InternalError, NameSpace
 from psyclone.psyir.transformations import TransformationError
 from psyclone.psyir.transformations import ProfileTrans
 from psyclone.tests.utilities import get_invoke
@@ -97,7 +98,7 @@ def test_profile_node_invalid_name(value):
 def test_profile_basic(capsys):
     '''Check basic functionality: node names, schedule view.
     '''
-    from psyclone.psyGen import colored, SCHEDULE_COLOUR_MAP
+    from psyclone.psyir.nodes.node import colored, SCHEDULE_COLOUR_MAP
     Profiler.set_options([Profiler.INVOKES])
     _, invoke = get_invoke("test11_different_iterates_over_one_invoke.f90",
                            "gocean1.0", idx=0)
@@ -632,7 +633,7 @@ End Schedule""")
     sched3.view()
     out, _ = capsys.readouterr()
 
-    from psyclone.psyGen import SCHEDULE_COLOUR_MAP, colored
+    from psyclone.psyir.nodes.node import SCHEDULE_COLOUR_MAP, colored
     gsched = colored("GOInvokeSchedule", SCHEDULE_COLOUR_MAP["Schedule"])
     prof = colored("Profile", SCHEDULE_COLOUR_MAP["Profile"])
     sched = colored("Schedule", SCHEDULE_COLOUR_MAP["Schedule"])
