+++ src/psyclone/tests/psyir/transformations/transformations_test.py
@@ -40,7 +40,7 @@ API-agnostic tests for various transformation classes.
 
 from __future__ import absolute_import, print_function
 import pytest
-from psyclone.psyGen import Literal, Loop, Node, Reference, Schedule
+from psyclone.psyir.nodes import Literal, Loop, Node, Reference, Schedule
 from psyclone.psyir.symbols import DataType
 from psyclone.psyir.transformations import ProfileTrans, RegionTrans, \
     TransformationError
@@ -106,7 +106,7 @@ def test_ifblock_children_region():
     ''' Check that we reject attempts to transform the conditional part of
     an If statement or to include both the if- and else-clauses in a region
     (without their parent). '''
-    from psyclone.psyGen import IfBlock
+    from psyclone.psyir.nodes import IfBlock
     acct = ACCParallelTrans()
     # Construct a valid IfBlock
     ifblock = IfBlock()
@@ -134,7 +134,7 @@ def test_ifblock_children_region():
 
 def test_fusetrans_error_incomplete():
     ''' Check that we reject attempts to fuse loops which are incomplete. '''
-    from psyclone.psyGen import Return
+    from psyclone.psyir.nodes import Return
     from psyclone.transformations import LoopFuseTrans
     sch = Schedule()
     loop1 = Loop(variable_name="i", parent=sch)
